+++ src/blob/index/core.rs
@@ -44,7 +44,6 @@ pub(crate) struct IndexStruct<FileIndex, K>
 where
     for<'a> K: Key<'a>,
 {
-    mem: Option<MemoryAttrs>,
     filter: CombinedFilter<K>,
     bloom_offset: Option<u64>,
     params: IndexParams,
@@ -54,21 +53,74 @@ where
 }
 
 #[derive(Debug, Default)] // Default can be used to initialize structure with 0
-pub(crate) struct MemoryAttrs {
-    pub(crate) key_size: usize,
-    pub(crate) btree_entry_size: usize,
-    pub(crate) records_allocated: usize,
-    // contains actual size occupied by record header in RAM (which helps
-    // to compute actual size of indices in RAM in `InMemory` state)
-    pub(crate) record_header_size: usize,
-    pub(crate) records_count: usize,
+struct MemoryAttrs<K> {
+    records_count: usize,
+    records_allocated: usize,
+    marker: PhantomData<K>,
+}
+
+impl<K> MemoryAttrs<K>
+where
+    for<'a> K: Key<'a>,
+{
+    const KEY_SIZE: usize = K::LEN as usize;
+    const BPTREE_ENTRY_SIZE: usize =
+        size_of::<Vec<u8>>() + MemoryAttrs::<K>::KEY_SIZE + size_of::<Vec<RecordHeader>>();
+    const RECORD_HEADER_SIZE: usize = size_of::<RecordHeader>() + MemoryAttrs::<K>::KEY_SIZE;
 }
 
 pub type InMemoryIndex<K> = BTreeMap<K, Vec<RecordHeader>>;
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Default)]
+pub(crate) struct InMemoryData<K> {
+    headers: InMemoryIndex<K>,
+    mem: MemoryAttrs<K>,
+}
+
+impl<K> InMemoryData<K>
+where
+    for<'a> K: Key<'a>,
+{
+    fn new(headers: InMemoryIndex<K>, count: usize) -> Self {
+        let mem = MemoryAttrs {
+            records_allocated: headers.values().fold(0, |acc, v| acc + v.capacity()),
+            records_count: count,
+            marker: PhantomData,
+        };
+
+        Self { headers, mem }
+    }
+
+    fn memory_used(&self) -> usize {
+        let Self { mem, .. } = &self;
+        let MemoryAttrs {
+            records_count,
+            records_allocated,
+            ..
+        } = &mem;
+        let len = self.headers.len();
+        trace!("record_header_size: {}, records_allocated: {}, data.len(): {}, entry_size (key + vec): {}",
+                MemoryAttrs::<K>::RECORD_HEADER_SIZE, records_allocated, len, MemoryAttrs::<K>::BPTREE_ENTRY_SIZE);
+        // last minus is neccessary, because allocated but not initialized record
+        // headers don't have key allocated on heap
+        MemoryAttrs::<K>::RECORD_HEADER_SIZE * records_allocated
+            + len * MemoryAttrs::<K>::BPTREE_ENTRY_SIZE
+            - (records_allocated - records_count) * MemoryAttrs::<K>::KEY_SIZE
+    }
+
+    fn records_count(&self) -> usize {
+        self.mem.records_count
+    }
+
+    fn register_record_allocation(&mut self, records_allocated: usize) {
+        self.mem.records_allocated += records_allocated;
+        self.mem.records_count += 1;
+    }
+}
+
+#[derive(Debug)]
 pub(crate) enum State<FileIndex, K> {
-    InMemory(InMemoryIndex<K>),
+    InMemory(SRwLock<InMemoryData<K>>),
     OnDisk(FileIndex),
 }
 
@@ -80,21 +132,18 @@ where
     pub(crate) fn new(name: FileName, iodriver: IoDriver, config: IndexConfig) -> Self {
         let params = IndexParams::new(config.bloom_config.is_some(), config.recreate_index_file);
         let bloom_filter = config.bloom_config.map(|cfg| Bloom::new(cfg));
-        let mem = Some(Default::default());
         Self {
             params,
             filter: CombinedFilter::new(bloom_filter, RangeFilter::new()),
             bloom_offset: None,
-            inner: State::InMemory(BTreeMap::new()),
-            mem,
+            inner: State::InMemory(SRwLock::default()),
             name,
             iodriver,
         }
     }
 
     pub(crate) fn clear(&mut self) {
-        self.inner = State::InMemory(BTreeMap::new());
-        self.mem = Some(Default::default());
+        self.inner = State::InMemory(SRwLock::default());
         self.filter.clear_filter();
     }
 
@@ -135,7 +184,6 @@ where
         trace!("index restored successfuly");
         let index = Self {
             inner: State::OnDisk(findex),
-            mem: None,
             name,
             filter: CombinedFilter::new(bloom_filter, range_filter),
             bloom_offset: Some(bloom_offset as u64),
@@ -151,6 +199,10 @@ where
 
     async fn dump_in_memory(&mut self, blob_size: u64) -> Result<usize> {
         if let State::InMemory(headers) = &self.inner {
+            let headers = {
+                let mut headers = headers.write().expect("rwlock");
+                std::mem::take(&mut *headers).headers
+            };
             if headers.len() == 0 {
                 return Ok(0);
             }
@@ -160,7 +212,7 @@ where
             let findex = FileIndex::from_records(
                 &self.name.to_path(),
                 self.iodriver.clone(),
-                headers,
+                &headers,
                 meta_buf,
                 self.params.recreate_file,
                 blob_size,
@@ -168,11 +220,9 @@ where
             .await?;
             let size = findex.file_size() as usize;
             self.inner = State::OnDisk(findex);
-            self.mem = None;
-            Ok(size)
-        } else {
-            Ok(0)
+            return Ok(size);
         }
+        Ok(0)
     }
 
     fn serialize_filters(&self) -> Result<(Vec<u8>, usize)> {
@@ -203,8 +253,7 @@ where
 
     async fn load_in_memory(&mut self, findex: FileIndex, blob_size: u64) -> Result<()> {
         let (record_headers, records_count) = findex.get_records_headers(blob_size).await?;
-        self.mem = Some(compute_mem_attrs(&record_headers, records_count));
-        self.inner = State::InMemory(record_headers);
+        self.inner = State::InMemory(SRwLock::new(InMemoryData::new(record_headers, records_count)));
         let meta_buf = findex.read_meta().await?;
         let (bloom_filter, range_filter, _) = Self::deserialize_filters(&meta_buf)?;
         let bloom_filter = if self.params.bloom_is_on {
@@ -218,20 +267,9 @@ where
     }
 
     pub(crate) fn memory_used(&self) -> usize {
-        if let State::InMemory(data) = &self.inner {
-            let mem = self
-                .mem
-                .as_ref()
-                .expect("No memory info in `InMemory` State");
-            trace!("record_header_size: {}, records_allocated: {}, data.len(): {}, entry_size (key + vec): {}",
-            mem.record_header_size, mem.records_allocated, data.len(), mem.btree_entry_size
-            );
-            // last minus is neccessary, because allocated but not initialized record headers don't
-            // have key allocated on heap
-            mem.record_header_size * mem.records_allocated + data.len() * mem.btree_entry_size
-                - (mem.records_allocated - mem.records_count) * mem.key_size
-        } else {
-            0
+        match &self.inner {
+            State::InMemory(data) => data.read().expect("rwlock").memory_used(),
+            State::OnDisk(file) => file.memory_used(),
         }
     }
 
@@ -256,31 +294,26 @@ where
             .map(|h| h.map(|h| BlobRecordTimestamp::new(h.created())))
     }
 
-    fn push(&mut self, key: &K, h: RecordHeader) -> Result<()> {
+    fn push(&self, key: &K, h: RecordHeader) -> Result<()> {
         debug!("blob index simple push");
-        match &mut self.inner {
+        match &self.inner {
             State::InMemory(headers) => {
+                let mut data = headers.write().expect("rwlock");
                 debug!("blob index simple push bloom filter add");
                 self.filter.add(key);
                 debug!("blob index simple push key: {:?}", h.key());
-                let mem = self
-                    .mem
-                    .as_mut()
-                    .expect("No memory info in `InMemory` State");
-                if let Some(v) = headers.get_mut(key) {
+                let records_allocated;
+                if let Some(v) = data.headers.get_mut(key) {
                     let old_capacity = v.capacity();
                     v.push(h);
                     trace!("capacity growth: {}", v.capacity() - old_capacity);
-                    mem.records_allocated += v.capacity() - old_capacity;
+                    records_allocated = v.capacity() - old_capacity;
                 } else {
-                    if mem.records_count == 0 {
-                        set_key_related_fields::<K>(mem);
-                    }
                     let v = vec![h];
-                    mem.records_allocated += v.capacity(); // capacity == 1
-                    headers.insert(key.clone(), v);
+                    records_allocated = v.capacity(); // capacity == 1
+                    data.headers.insert(key.clone(), v);
                 }
-                mem.records_count += 1;
+                data.register_record_allocation(records_allocated);
                 Ok(())
             }
             State::OnDisk(_) => Err(Error::from(ErrorKind::Index(
@@ -302,12 +335,15 @@ where
 
     async fn get_all_with_deletion_marker(&self, key: &K) -> Result<Vec<RecordHeader>> {
         let headers = match &self.inner {
-            State::InMemory(headers) => Ok(headers.get(key).cloned().map(|mut hs| {
-                if hs.len() > 1 {
-                    hs.reverse();
-                }
-                hs
-            })),
+            State::InMemory(data) => {
+                let data = data.read().expect("rwlock");
+                Ok(data.headers.get(key).cloned().map(|mut hs| {
+                    if hs.len() > 1 {
+                        hs.reverse();
+                    }
+                    hs
+                }))
+            }
             State::OnDisk(findex) => findex.find_by_key(key).await,
         }?;
         if let Some(mut hs) = headers {
@@ -325,10 +361,11 @@ where
         debug!("index get any");
         let result = match &self.inner {
             State::InMemory(headers) => {
-                debug!("index get any in memory headers: {}", headers.len());
+                let data = headers.read().expect("rwlock");
+                debug!("index get any in memory headers: {}", data.headers.len());
                 // in memory indexes with same key are stored in ascending order, so the last
                 // by adding time record is last in list (in b+tree disk index it's first)
-                headers.get(key).and_then(|h| h.last()).cloned()
+                data.headers.get(key).and_then(|h| h.last()).cloned()
             }
             State::OnDisk(findex) => {
                 debug!("index get any on disk");
@@ -359,13 +396,9 @@ where
     }
 
     fn count(&self) -> usize {
-        match self.inner {
+        match &self.inner {
             State::OnDisk(ref findex) => findex.records_count(),
-            State::InMemory(_) => self
-                .mem
-                .as_ref()
-                .map(|mem| mem.records_count)
-                .expect("No memory info in `InMemory` State"),
+            State::InMemory(d) => d.read().expect("rwlock").records_count(),
         }
     }
 
@@ -397,6 +430,7 @@ pub(crate) trait FileIndexTrait<K>: Sized + Send + Sync {
     async fn get_records_headers(&self, blob_size: u64) -> Result<(InMemoryIndex<K>, usize)>;
     async fn get_any(&self, key: &K) -> Result<Option<RecordHeader>>;
     fn validate(&self, blob_size: u64) -> Result<()>;
+    fn memory_used(&self) -> usize;
 }
 
 #[async_trait::async_trait]